検索機能の追加

+++ handler/posts.go
@@ -49,7 +49,6 @@ func CreatePost(w http.ResponseWriter, r *http.Request) {
 		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.InvalidContentType)
 		return
 	}
-	
 
 	stmt, err := db.Db.Prepare("INSERT INTO posts(title, user_id, created_at, updated_at) VALUES(?, ?, ?, ?)")
 	if err != nil {
@@ -68,6 +67,17 @@ func CreatePost(w http.ResponseWriter, r *http.Request) {
 	createdAt := time.Now()
 	updatedAt := time.Now()
 
+	if title == "" || userID == 0 {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.TitleOrIdIsEmpty)
+		return
+	}
+
+	// POSTで送られてきたuser_idのuserが存在するかチェック
+	if !utils.CheckUserExist(keyVal["user_id"]) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.UserIdNotFound)
+		return
+	}
+
 	_, err = stmt.Exec(title, userID, createdAt, updatedAt)
 	if err != nil {
 		log.Println(fmt.Errorf("error at stmt.Exec: %w", err))
@@ -79,6 +89,14 @@ func CreatePost(w http.ResponseWriter, r *http.Request) {
 
 func GetPost(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
+
+	fmt.Printf("%T\n", params)
+
+	// パラメータに指定されたpostのIDが存在するかのチェック
+	if !utils.CheckPostExist(params["id"]) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.PostNotFound)
+		return
+	}
 
 	result, err := db.Db.Query("SELECT id, title, user_id, created_at, updated_at FROM posts WHERE id = ?", params["id"])
 	if err != nil {
@@ -95,7 +113,7 @@ func GetPost(w http.ResponseWriter, r *http.Request) {
 		}
 
 	}
-	
+
 	postJSON, err := json.Marshal(post)
 	if err != nil {
 		log.Println(fmt.Errorf("error at JSONMarshal: %w", err))
@@ -112,6 +130,12 @@ func UpdatePost(w http.ResponseWriter, r *http.Request) {
 
 	params := mux.Vars(r)
 
+	// パラメータに指定されたIDのpostが存在するかのチェック
+	if !utils.CheckPostExist(params["id"]) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.PostNotFound)
+		return
+	}
+
 	stmt, err := db.Db.Prepare("UPDATE posts SET title = ?, updated_at = ? WHERE id = ?")
 	if err != nil {
 		log.Println(fmt.Errorf("error at Query: %w", err))
@@ -125,6 +149,10 @@ func UpdatePost(w http.ResponseWriter, r *http.Request) {
 	keyVal := make(map[string]string)
 	json.Unmarshal(body, &keyVal)
 	newTitle := keyVal["title"]
+	if newTitle == "" {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.TitleIsEmpty)
+		return
+	}
 	updatedAt := time.Now()
 
 	_, err = stmt.Exec(newTitle, updatedAt, params["id"])
@@ -137,6 +165,11 @@ func UpdatePost(w http.ResponseWriter, r *http.Request) {
 
 func DeletePost(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
+
+	if !utils.CheckPostExist(params["id"]) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.PostNotFound)
+		return
+	}
 
 	stmt, err := db.Db.Prepare("DELETE FROM posts WHERE id = ?")
 	if err != nil {
