+++ handler/users.go
@@ -28,7 +28,7 @@ func GetUsers(w http.ResponseWriter, r *http.Request) {
 	for result.Next() {
 		var user models.User
 
-		err := result.Scan(&user.ID, &user.Name, &user.CreatedAt, &user.UpdatedAt)
+		err := result.Scan(&user.ID, &user.Name, &user.Email, &user.CreatedAt, &user.UpdatedAt)
 		if err != nil {
 			log.Println(fmt.Errorf("error at Scan: %w", err))
 		}
@@ -45,12 +45,12 @@ func GetUsers(w http.ResponseWriter, r *http.Request) {
 }
 
 func CreateUser(w http.ResponseWriter, r *http.Request) {
-	if contentType := r.Header.Get("Content-Type"); contentType != "application/json"{
+	if contentType := r.Header.Get("Content-Type"); contentType != "application/json" {
 		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.InvalidContentType)
 		return
 	}
 
-	stmt, err := db.Db.Prepare("INSERT INTO users(name, created_at, updated_at) VALUES(?, ?, ?)")
+	stmt, err := db.Db.Prepare("INSERT INTO users(name, email, created_at, updated_at) VALUES(?, ?, ?, ?)")
 	if err != nil {
 		log.Println(fmt.Errorf("error at Query: %w", err))
 	}
@@ -63,9 +63,22 @@ func CreateUser(w http.ResponseWriter, r *http.Request) {
 	keyVal := make(map[string]string)
 	json.Unmarshal(body, &keyVal)
 	name := keyVal["name"]
+	email := keyVal["email"]
 	createdAt := time.Now()
 	updatedAt := time.Now()
-	_, err = stmt.Exec(name, createdAt, updatedAt)
+
+	if name == "" || email == "" {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.NameOrEmailIsEmpty)
+		return
+	}
+
+	// ユーザー作成時のemailが重複していないかの確認
+	if !utils.CheckEmailExist(email) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.InvalidEmail)
+		return
+	}
+
+	_, err = stmt.Exec(name, email, createdAt, updatedAt)
 	if err != nil {
 		log.Println(fmt.Errorf("error at stmt.Exec: %w", err))
 	}
@@ -76,7 +89,14 @@ func CreateUser(w http.ResponseWriter, r *http.Request) {
 
 func GetUser(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
-	result, err := db.Db.Query("SELECT id, name, created_at, updated_at FROM users WHERE id = ?", params["id"])
+
+	// パラメータで指定されたIDのuserが存在するかチェック
+	if !utils.CheckUserExist(params["id"]) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.UserNotFound)
+		return
+	}
+
+	result, err := db.Db.Query("SELECT id, name, email, created_at, updated_at FROM users WHERE id = ?", params["id"])
 	if err != nil {
 		log.Println(fmt.Errorf("error at Query: %w", err))
 	}
@@ -84,7 +104,7 @@ func GetUser(w http.ResponseWriter, r *http.Request) {
 	defer result.Close()
 	var user models.User
 	for result.Next() {
-		err := result.Scan(&user.ID, &user.Name, &user.CreatedAt, &user.UpdatedAt)
+		err := result.Scan(&user.ID, &user.Name, &user.Email, &user.CreatedAt, &user.UpdatedAt)
 		if err != nil {
 			log.Println(fmt.Errorf("error at Scan: %w", err))
 		}
@@ -106,7 +126,13 @@ func UpdateUser(w http.ResponseWriter, r *http.Request) {
 
 	params := mux.Vars(r)
 
-	stmt, err := db.Db.Prepare("UPDATE users SET name = ?, updated_at = ? WHERE id = ?")
+	// パラメータで指定されたuserIDが存在するかの確認
+	if !utils.CheckUserExist(params["id"]) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.UserIdNotFound)
+		return
+	}
+
+	stmt, err := db.Db.Prepare("UPDATE users SET name = ?, email = ?, updated_at = ? WHERE id = ?")
 	if err != nil {
 		log.Println(fmt.Errorf("error at Query: %w", err))
 	}
@@ -118,9 +144,37 @@ func UpdateUser(w http.ResponseWriter, r *http.Request) {
 	keyVal := make(map[string]string)
 	json.Unmarshal(body, &keyVal)
 	name := keyVal["name"]
+	email := keyVal["email"]
 	updatedAt := time.Now()
 
-	_, err = stmt.Exec(name, updatedAt, params["id"])
+	if name == "" && email == "" {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.NameAndEmailIsEmpty)
+		return
+	} else if name == "" {
+		stmt, err := db.Db.Prepare("UPDATE users SET email = ?, updated_at = ? WHERE id = ?")
+		if err != nil {
+			log.Println(fmt.Errorf("error at Query1: %w", err))
+		}
+		_, err = stmt.Exec(email, updatedAt, params["id"])
+		if err != nil {
+			log.Println(fmt.Errorf("error at stmt.Exec: %w", err))
+		}
+		utils.ReturnJsonResponse(w, http.StatusOK, utils.UserUpdateSuccess)
+		return
+	} else if email == "" {
+		stmt, err := db.Db.Prepare("UPDATE users SET name = ?, updated_at = ? WHERE id = ?")
+		if err != nil {
+			log.Println(fmt.Errorf("error at Query2: %w", err))
+		}
+		_, err = stmt.Exec(name, updatedAt, params["id"])
+		if err != nil {
+			log.Println(fmt.Errorf("error at stmt.Exec: %w", err))
+		}
+		utils.ReturnJsonResponse(w, http.StatusOK, utils.UserUpdateSuccess)
+		return
+	}
+
+	_, err = stmt.Exec(name, email, updatedAt, params["id"])
 	if err != nil {
 		log.Println(fmt.Errorf("error at stmt.Exec: %w", err))
 	}
@@ -130,6 +184,12 @@ func UpdateUser(w http.ResponseWriter, r *http.Request) {
 
 func DeleteUser(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
+
+	// パラメータで指定されたuserIDが存在するかの確認
+	if !utils.CheckUserExist(params["id"]) {
+		utils.ReturnJsonResponse(w, http.StatusBadRequest, utils.UserIdNotFound)
+		return
+	}
 
 	stmt, err := db.Db.Prepare("DELETE FROM users WHERE id = ?")
 	if err != nil {
